requirements.txt
absl-py==1.4.0
aniso8601==8.0.0
astor==0.8.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
gast==0.5.4
greenlet==2.0.2
grpcio==1.56.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
marshmallow==3.5.2
marshmallow-sqlalchemy==0.23.0
mock==5.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
PyJWT==2.4.0
pymongo==4.4.1
PyMySQL==1.1.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
regex==2023.6.3
requests==2.31.0
six==1.14.0
SQLAlchemy==1.3.16
tensorflow-estimator==1.13.0
termcolor==2.3.0
tflearn==0.5.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
